refactor: custom-hooks moves logic to custom hooks
refactor: custom hooks adds absolute path
