+++ src/components/index.ts
@@ -1,4 +1,4 @@
 import Operators from './Operators/Operators';
 import OperatorsTable from './OperatorsTable/OperatorsTable';
 
-export { Operators, OperatorsTable }; //todo: think more
+export { Operators, OperatorsTable };
