+++ src/components/shared/ErrorMessage/ErrorMessage.tsx
@@ -1,6 +1,6 @@
 import { Typography } from '@mui/material';
+import theme from 'theme';
 import { StyledContainer } from '../styles';
-import theme from '../../../theme';
 
 interface Props {
     error: string;
