+++ src/components/shared/SeachBox/SearchBox.tsx
@@ -1,5 +1,5 @@
 import { TextField } from '@mui/material';
-import React, { ChangeEvent } from 'react';
+import React from 'react';
 
 interface Props {
     value: string;
