+++ src/components/shared/index.ts
@@ -1,5 +1,6 @@
 import ErrorMessage from './ErrorMessage/ErrorMessage';
 import Loader from './Loader/Loader';
+import SearchBox from './SeachBox/SearchBox';
 import UserCard from './UserCard/UserCard';
 
-export { Loader, ErrorMessage, UserCard };
+export { Loader, ErrorMessage, UserCard, SearchBox };
