+++ src/pages/WelcomePage.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Button, Container, Typography } from '@mui/material';
-import { UniTalkRoute } from '../constants';
 import { Link } from 'react-router-dom';
+import { UniTalkRoute } from 'constants';
 
 const WelcomePage: React.FC = () => {
     return (
