+++ src/pages/index.ts
@@ -1,0 +1,4 @@
+import OperatorsPage from './OperatorsPage';
+import WelcomePage from './WelcomePage';
+
+export { OperatorsPage, WelcomePage };
