+++ src/reducers/index.ts
@@ -1,8 +1,8 @@
 import { combineReducers } from '@reduxjs/toolkit';
-import operatorReducer from '../slices/operatorSlice';
+import operatorReducer from 'slices/operatorSlice';
 
 const rootReducer = combineReducers({
-    operator: operatorReducer,
+    operator: operatorReducer, //todo: only 1 reducer
 });
 
 export default rootReducer;
