+++ src/services/index.ts
@@ -1,6 +1,6 @@
 import axios from 'axios';
 import { BASE_URL } from './config';
-import { Operator, OperatorAddon } from '../types';
+import { Operator, OperatorAddon } from 'types';
 
 const api = axios.create({
     baseURL: BASE_URL,
