+++ tsconfig.json
@@ -7,8 +7,8 @@
         "moduleResolution": "node",
         "esModuleInterop": true,
         "skipLibCheck": true,
-        "forceConsistentCasingInFileNames": true
-        //todo: add base url
+        "forceConsistentCasingInFileNames": true,
+        "baseUrl": "src"
     },
     "include": ["src"]
 }
